:zap: 비즈니스계정 버젼 재정리 시작

+++ src/main/java/com/adplatform/restApi/batch/dto/BatchStatusDto.java
@@ -33,7 +33,7 @@ public class BatchStatusDto {
             private Integer companyId;
             private Integer ownerCompanyId;
             private Integer reportDate;
-            private Long cost;
+            private Float cost;
         }
 
         @Getter
